refactor: adjusing protectedRoute structure
