+++ client/src/context/AuthContext.tsx
@@ -1,43 +1,9 @@
-import { createContext, useContext, useState, useEffect } from "react";
+import { createContext } from "react";
 
 interface AuthContextData {
-  isAuthenticated: boolean;
-  login: () => void;
-  logout: () => void;
+    isAuthenticated: boolean;
+    login: () => void;
+    logout: () => void;
 }
 
-const AuthContext = createContext<AuthContextData | undefined>(undefined);
-
-export const AuthProvider = ({ children }: { children: React.ReactNode }) => {
-  const [isAuthenticated, setIsAuthenticated] = useState<boolean>(false);
-
-  useEffect(() => {
-    const token = localStorage.getItem("authToken");
-    if (token) {
-      setIsAuthenticated(true);
-    }
-  }, []);
-
-  const login = () => {
-    setIsAuthenticated(true);
-  };
-  
-  const logout = () => {
-    setIsAuthenticated(false);
-    localStorage.removeItem("authToken");
-  };
-
-  return (
-    <AuthContext.Provider value={{ isAuthenticated, login, logout }}>
-      {children}
-    </AuthContext.Provider>
-  );
-};
-
-export const useAuth = () => {
-  const context = useContext(AuthContext);
-  if (!context) {
-    throw new Error("useAuth must be used within an AuthProvider");
-  }
-  return context;
-};
+export const AuthContext = createContext<AuthContextData | undefined>(undefined);
