+++ client/src/context/AuthProvider.tsx
@@ -1,0 +1,26 @@
+import { useEffect, useState } from "react"
+import { AuthContext } from "./AuthContext";
+
+export const AuthProvider = ({ children }: { children: React.ReactNode }) => {
+    const [isAuthenticated, setIsAuthenticated] = useState<boolean>(false);
+
+    useEffect(() => {
+        const token = localStorage.getItem("authToken");
+        if (token) setIsAuthenticated(true);
+    }, []);
+
+    const login = () => {
+        setIsAuthenticated(true);
+    };
+
+    const logout = () => {
+        setIsAuthenticated(false);
+        localStorage.removeItem("authToken");
+    };
+
+    return (
+        <AuthContext.Provider value={{ isAuthenticated, login, logout }}>
+            {children}
+        </AuthContext.Provider>
+    );
+}
